requirements.txt
alembic==1.5.4
autopep8==1.5.5
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycodestyle==2.6.0
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
twilio==6.52.0
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
